+++ src/thinpi/tpupdate/tpupdate.x.c
@@ -1,0 +1,716 @@
+#if 0
+	shc Version 4.0.3, Generic Shell Script Compiler
+	GNU GPL Version 3 Md Jahidul Hamid <jahidulhamid@yahoo.com>
+
+	shc -f src/thinpi/tpupdate/tpupdate 
+#endif
+
+static  char data [] = 
+#define      msg1_z	65
+#define      msg1	((&data[10]))
+	"\053\114\165\172\003\355\305\256\254\377\203\265\136\157\255\172"
+	"\047\252\025\260\322\335\117\010\140\177\172\125\015\276\123\060"
+	"\077\245\126\207\323\221\260\263\245\265\006\270\205\203\013\336"
+	"\255\342\101\307\105\140\014\127\350\314\250\202\206\055\150\052"
+	"\341\220\164\010\100\222\242\106\136\270\275\152\122\351\142\072"
+	"\321\105"
+#define      date_z	1
+#define      date	((&data[82]))
+	"\313"
+#define      shll_z	10
+#define      shll	((&data[85]))
+	"\014\251\257\324\326\222\236\341\261\374\066\343\352\033"
+#define      rlax_z	1
+#define      rlax	((&data[97]))
+	"\373"
+#define      inlo_z	3
+#define      inlo	((&data[98]))
+	"\252\372\110"
+#define      msg2_z	19
+#define      msg2	((&data[104]))
+	"\207\157\267\363\366\367\227\132\037\337\222\371\012\342\037\270"
+	"\170\371\006\121\371\313"
+#define      xecc_z	15
+#define      xecc	((&data[124]))
+	"\206\205\002\114\173\123\230\340\042\276\054\306\053\037\233\023"
+#define      lsto_z	1
+#define      lsto	((&data[139]))
+	"\071"
+#define      tst1_z	22
+#define      tst1	((&data[144]))
+	"\352\163\004\247\156\042\221\312\336\016\333\111\164\060\244\141"
+	"\252\244\131\322\234\244\226\142\171\354\040\020\120\013\053"
+#define      chk2_z	19
+#define      chk2	((&data[171]))
+	"\276\101\210\053\355\306\036\357\234\010\077\115\120\135\102\302"
+	"\130\310\031\041\075\352\274"
+#define      tst2_z	19
+#define      tst2	((&data[198]))
+	"\126\252\170\305\352\217\040\020\077\202\344\127\015\341\116\355"
+	"\241\117\227\364\071\120\125\314"
+#define      pswd_z	256
+#define      pswd	((&data[225]))
+	"\300\107\033\233\061\216\240\136\312\262\063\274\314\100\160\131"
+	"\153\231\016\254\225\035\102\244\152\073\113\332\371\024\170\041"
+	"\161\337\302\270\156\116\027\070\000\112\365\314\212\146\046\365"
+	"\000\064\241\225\121\344\072\274\040\206\227\031\232\020\073\014"
+	"\357\376\304\135\114\333\226\115\046\214\032\260\363\100\246\363"
+	"\165\110\210\307\054\302\204\114\110\033\146\343\053\242\357\032"
+	"\240\264\170\354\220\017\072\266\233\124\147\216\225\015\201\013"
+	"\125\012\322\202\315\127\316\025\162\065\370\236\327\350\271\167"
+	"\234\062\144\055\101\237\343\335\363\112\153\211\130\355\224\256"
+	"\370\147\060\305\276\377\332\061\064\323\320\013\273\212\203\130"
+	"\274\350\205\375\207\151\332\173\264\106\005\014\064\232\272\054"
+	"\001\352\361\300\351\314\362\036\237\302\052\133\114\256\264\011"
+	"\226\071\006\036\243\341\232\127\050\237\144\134\071\036\211\073"
+	"\011\172\373\363\106\356\021\346\260\074\102\375\352\366\006\201"
+	"\060\015\237\323\357\072\053\027\332\217\164\023\256\375\117\267"
+	"\170\113\252\277\071\274\245\352\370\350\350\342\336\356\143\017"
+	"\374\003\343\353\075\016\003\330\257\260\051\272\334\046\310\301"
+	"\110\006\254\004\354\237\110\103\112\300\010"
+#define      chk1_z	22
+#define      chk1	((&data[505]))
+	"\150\200\154\203\362\035\132\271\203\250\257\010\135\121\032\206"
+	"\306\116\344\167\104\055\312\351\303\021"
+#define      opts_z	1
+#define      opts	((&data[527]))
+	"\142"
+#define      text_z	393
+#define      text	((&data[574]))
+	"\155\240\174\112\307\104\014\017\113\270\024\067\130\134\172\243"
+	"\035\203\272\334\250\042\135\025\246\172\264\270\066\053\172\244"
+	"\314\366\356\224\073\372\244\206\263\270\275\013\025\070\020\145"
+	"\027\054\007\143\314\061\150\265\305\054\033\361\015\301\072\072"
+	"\242\074\352\340\173\107\205\321\273\272\004\074\007\071\132\217"
+	"\346\237\110\357\157\137\170\232\324\033\175\365\063\121\243\322"
+	"\206\312\101\273\345\177\243\330\366\075\030\060\204\355\131\274"
+	"\102\176\173\250\132\011\147\202\365\173\054\237\120\075\327\206"
+	"\172\204\227\110\320\311\241\063\107\101\274\310\100\262\371\230"
+	"\364\006\302\241\003\065\077\000\016\263\044\226\275\042\171\172"
+	"\042\042\277\171\040\215\175\307\164\363\057\177\077\204\207\212"
+	"\353\167\167\167\346\203\173\204\215\233\362\176\225\001\346\354"
+	"\306\134\165\271\122\360\372\215\251\320\320\371\163\330\225\362"
+	"\143\161\254\270\126\353\122\217\102\213\357\346\053\263\231\355"
+	"\031\132\035\053\237\031\073\020\247\022\223\223\215\217\106\161"
+	"\233\231\320\011\036\353\266\202\317\355\044\041\333\121\124\350"
+	"\262\354\360\337\067\376\343\134\065\141\113\011\071\134\204\034"
+	"\101\030\115\236\316\307\340\236\303\337\202\020\240\033\306\155"
+	"\003\011\366\326\014\222\303\271\320\161\340\076\076\125\354\212"
+	"\145\142\113\117\301\060\160\256\251\265\124\260\210\067\325\016"
+	"\030\174\354\273\355\047\145\040\331\375\321\352\042\145\141\257"
+	"\016\060\315\034\330\141\131\002\061\031\343\124\302\372\016\133"
+	"\022\116\076\105\152\047\322\332\124\343\000\317\051\015\312\214"
+	"\001\111\063\307\172\320\357\021\155\143\276\353\173\372\142\025"
+	"\257\360\044\022\255\063\035\035\101\152\074\000\317\041\347\267"
+	"\132\021\017\246\235\234\024\125\003\315\117\302\037\323\170\304"
+	"\322\135\231\112\252\326\164\215\244\302\072\175\041\043\164\174"
+	"\246\257\047\236\007\270\023\256\062\274\150\017\145\213\154\173"
+	"\061\346\057\352\035\133\144\301\047\132\260\274\226\253\140\034"
+	"\136\030\332\151\056\023\030\140\317\200\160\064\014\335\257\076"
+	"\304\336\050\341\072\215\243\141\350\124\035\176\377\176\232\135"
+	"\227\165\307\305\210\337\046\127\137\226\214\154\163\073\252\067"
+	"\032"/* End of data[] */;
+#define      hide_z	4096
+#define SETUID 0	/* Define as 1 to call setuid(0) at start of script */
+#define DEBUGEXEC	0	/* Define as 1 to debug execvp calls */
+#define TRACEABLE	1	/* Define as 1 to enable ptrace the executable */
+#define HARDENING	0	/* Define as 1 to disable ptrace/dump the executable */
+#define BUSYBOXON	0	/* Define as 1 to enable work with busybox */
+
+#if HARDENING
+static const char * shc_x[] = {
+"/*",
+" * Copyright 2019 - Intika <intika@librefox.org>",
+" * Replace ******** with secret read from fd 21",
+" * Also change arguments location of sub commands (sh script commands)",
+" * gcc -Wall -fpic -shared -o shc_secret.so shc_secret.c -ldl",
+" */",
+"",
+"#define _GNU_SOURCE /* needed to get RTLD_NEXT defined in dlfcn.h */",
+"#define PLACEHOLDER \"********\"",
+"#include <dlfcn.h>",
+"#include <stdlib.h>",
+"#include <string.h>",
+"#include <unistd.h>",
+"#include <stdio.h>",
+"#include <signal.h>",
+"",
+"static char secret[128000]; //max size",
+"typedef int (*pfi)(int, char **, char **);",
+"static pfi real_main;",
+"",
+"// copy argv to new location",
+"char **copyargs(int argc, char** argv){",
+"    char **newargv = malloc((argc+1)*sizeof(*argv));",
+"    char *from,*to;",
+"    int i,len;",
+"",
+"    for(i = 0; i<argc; i++){",
+"        from = argv[i];",
+"        len = strlen(from)+1;",
+"        to = malloc(len);",
+"        memcpy(to,from,len);",
+"        // zap old argv space",
+"        memset(from,'\\0',len);",
+"        newargv[i] = to;",
+"        argv[i] = 0;",
+"    }",
+"    newargv[argc] = 0;",
+"    return newargv;",
+"}",
+"",
+"static int mymain(int argc, char** argv, char** env) {",
+"    //fprintf(stderr, \"Inject main argc = %d\\n\", argc);",
+"    return real_main(argc, copyargs(argc,argv), env);",
+"}",
+"",
+"int __libc_start_main(int (*main) (int, char**, char**),",
+"                      int argc,",
+"                      char **argv,",
+"                      void (*init) (void),",
+"                      void (*fini)(void),",
+"                      void (*rtld_fini)(void),",
+"                      void (*stack_end)){",
+"    static int (*real___libc_start_main)() = NULL;",
+"    int n;",
+"",
+"    if (!real___libc_start_main) {",
+"        real___libc_start_main = dlsym(RTLD_NEXT, \"__libc_start_main\");",
+"        if (!real___libc_start_main) abort();",
+"    }",
+"",
+"    n = read(21, secret, sizeof(secret));",
+"    if (n > 0) {",
+"      int i;",
+"",
+"    if (secret[n - 1] == '\\n') secret[--n] = '\\0';",
+"    for (i = 1; i < argc; i++)",
+"        if (strcmp(argv[i], PLACEHOLDER) == 0)",
+"          argv[i] = secret;",
+"    }",
+"",
+"    real_main = main;",
+"",
+"    return real___libc_start_main(mymain, argc, argv, init, fini, rtld_fini, stack_end);",
+"}",
+"",
+0};
+#endif /* HARDENING */
+
+/* rtc.c */
+
+#include <sys/stat.h>
+#include <sys/types.h>
+
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
+#include <unistd.h>
+
+/* 'Alleged RC4' */
+
+static unsigned char stte[256], indx, jndx, kndx;
+
+/*
+ * Reset arc4 stte. 
+ */
+void stte_0(void)
+{
+	indx = jndx = kndx = 0;
+	do {
+		stte[indx] = indx;
+	} while (++indx);
+}
+
+/*
+ * Set key. Can be used more than once. 
+ */
+void key(void * str, int len)
+{
+	unsigned char tmp, * ptr = (unsigned char *)str;
+	while (len > 0) {
+		do {
+			tmp = stte[indx];
+			kndx += tmp;
+			kndx += ptr[(int)indx % len];
+			stte[indx] = stte[kndx];
+			stte[kndx] = tmp;
+		} while (++indx);
+		ptr += 256;
+		len -= 256;
+	}
+}
+
+/*
+ * Crypt data. 
+ */
+void arc4(void * str, int len)
+{
+	unsigned char tmp, * ptr = (unsigned char *)str;
+	while (len > 0) {
+		indx++;
+		tmp = stte[indx];
+		jndx += tmp;
+		stte[indx] = stte[jndx];
+		stte[jndx] = tmp;
+		tmp += stte[indx];
+		*ptr ^= stte[tmp];
+		ptr++;
+		len--;
+	}
+}
+
+/* End of ARC4 */
+
+#if HARDENING
+
+#include <sys/ptrace.h>
+#include <sys/wait.h>
+#include <signal.h>
+#include <sys/prctl.h>
+#define PR_SET_PTRACER 0x59616d61
+
+/* Seccomp Sandboxing Init */
+#include <stdlib.h>
+#include <stdio.h>
+#include <stddef.h>
+#include <string.h>
+#include <unistd.h>
+#include <errno.h>
+
+#include <sys/types.h>
+#include <sys/prctl.h>
+#include <sys/syscall.h>
+#include <sys/socket.h>
+
+#include <linux/filter.h>
+#include <linux/seccomp.h>
+#include <linux/audit.h>
+
+#define ArchField offsetof(struct seccomp_data, arch)
+
+#define Allow(syscall) \
+    BPF_JUMP(BPF_JMP+BPF_JEQ+BPF_K, SYS_##syscall, 0, 1), \
+    BPF_STMT(BPF_RET+BPF_K, SECCOMP_RET_ALLOW)
+
+struct sock_filter filter[] = {
+    /* validate arch */
+    BPF_STMT(BPF_LD+BPF_W+BPF_ABS, ArchField),
+    BPF_JUMP( BPF_JMP+BPF_JEQ+BPF_K, AUDIT_ARCH_X86_64, 1, 0),
+    BPF_STMT(BPF_RET+BPF_K, SECCOMP_RET_KILL),
+
+    /* load syscall */
+    BPF_STMT(BPF_LD+BPF_W+BPF_ABS, offsetof(struct seccomp_data, nr)),
+
+    /* list of allowed syscalls */
+    Allow(exit_group),  /* exits a process */
+    Allow(brk),         /* for malloc(), inside libc */
+    Allow(mmap),        /* also for malloc() */
+    Allow(munmap),      /* for free(), inside libc */
+
+    /* and if we don't match above, die */
+    BPF_STMT(BPF_RET+BPF_K, SECCOMP_RET_KILL),
+};
+struct sock_fprog filterprog = {
+    .len = sizeof(filter)/sizeof(filter[0]),
+    .filter = filter
+};
+
+/* Seccomp Sandboxing - Set up the restricted environment */
+void seccomp_hardening() {
+    if (prctl(PR_SET_NO_NEW_PRIVS, 1, 0, 0, 0)) {
+        perror("Could not start seccomp:");
+        exit(1);
+    }
+    if (prctl(PR_SET_SECCOMP, SECCOMP_MODE_FILTER, &filterprog) == -1) {
+        perror("Could not start seccomp:");
+        exit(1);
+    }
+} 
+/* End Seccomp Sandboxing Init */
+
+void shc_x_file() {
+    FILE *fp;
+    int line = 0;
+
+    if ((fp = fopen("/tmp/shc_x.c", "w")) == NULL ) {exit(1); exit(1);}
+    for (line = 0; shc_x[line]; line++)	fprintf(fp, "%s\n", shc_x[line]);
+    fflush(fp);fclose(fp);
+}
+
+int make() {
+	char * cc, * cflags, * ldflags;
+    char cmd[4096];
+
+	cc = getenv("CC");
+	if (!cc) cc = "cc";
+
+	sprintf(cmd, "%s %s -o %s %s", cc, "-Wall -fpic -shared", "/tmp/shc_x.so", "/tmp/shc_x.c -ldl");
+	if (system(cmd)) {remove("/tmp/shc_x.c"); return -1;}
+	remove("/tmp/shc_x.c"); return 0;
+}
+
+void arc4_hardrun(void * str, int len) {
+    //Decode locally
+    char tmp2[len];
+    char tmp3[len+1024];
+    memcpy(tmp2, str, len);
+
+	unsigned char tmp, * ptr = (unsigned char *)tmp2;
+    int lentmp = len;
+    int pid, status;
+    pid = fork();
+
+    shc_x_file();
+    if (make()) {exit(1);}
+
+    setenv("LD_PRELOAD","/tmp/shc_x.so",1);
+
+    if(pid==0) {
+
+        //Start tracing to protect from dump & trace
+        if (ptrace(PTRACE_TRACEME, 0, 0, 0) < 0) {
+            kill(getpid(), SIGKILL);
+            _exit(1);
+        }
+
+        //Decode Bash
+        while (len > 0) {
+            indx++;
+            tmp = stte[indx];
+            jndx += tmp;
+            stte[indx] = stte[jndx];
+            stte[jndx] = tmp;
+            tmp += stte[indx];
+            *ptr ^= stte[tmp];
+            ptr++;
+            len--;
+        }
+
+        //Do the magic
+        sprintf(tmp3, "%s %s", "'********' 21<<<", tmp2);
+
+        //Exec bash script //fork execl with 'sh -c'
+        system(tmp2);
+
+        //Empty script variable
+        memcpy(tmp2, str, lentmp);
+
+        //Clean temp
+        remove("/tmp/shc_x.so");
+
+        //Sinal to detach ptrace
+        ptrace(PTRACE_DETACH, 0, 0, 0);
+        exit(0);
+    }
+    else {wait(&status);}
+
+    /* Seccomp Sandboxing - Start */
+    seccomp_hardening();
+
+    exit(0);
+}
+#endif /* HARDENING */
+
+/*
+ * Key with file invariants. 
+ */
+int key_with_file(char * file)
+{
+	struct stat statf[1];
+	struct stat control[1];
+
+	if (stat(file, statf) < 0)
+		return -1;
+
+	/* Turn on stable fields */
+	memset(control, 0, sizeof(control));
+	control->st_ino = statf->st_ino;
+	control->st_dev = statf->st_dev;
+	control->st_rdev = statf->st_rdev;
+	control->st_uid = statf->st_uid;
+	control->st_gid = statf->st_gid;
+	control->st_size = statf->st_size;
+	control->st_mtime = statf->st_mtime;
+	control->st_ctime = statf->st_ctime;
+	key(control, sizeof(control));
+	return 0;
+}
+
+#if DEBUGEXEC
+void debugexec(char * sh11, int argc, char ** argv)
+{
+	int i;
+	fprintf(stderr, "shll=%s\n", sh11 ? sh11 : "<null>");
+	fprintf(stderr, "argc=%d\n", argc);
+	if (!argv) {
+		fprintf(stderr, "argv=<null>\n");
+	} else { 
+		for (i = 0; i <= argc ; i++)
+			fprintf(stderr, "argv[%d]=%.60s\n", i, argv[i] ? argv[i] : "<null>");
+	}
+}
+#endif /* DEBUGEXEC */
+
+void rmarg(char ** argv, char * arg)
+{
+	for (; argv && *argv && *argv != arg; argv++);
+	for (; argv && *argv; argv++)
+		*argv = argv[1];
+}
+
+void chkenv_end(void);
+
+int chkenv(int argc)
+{
+	char buff[512];
+	unsigned long mask, m;
+	int l, a, c;
+	char * string;
+	extern char ** environ;
+
+	mask = (unsigned long)getpid();
+	stte_0();
+	 key(&chkenv, (void*)&chkenv_end - (void*)&chkenv);
+	 key(&data, sizeof(data));
+	 key(&mask, sizeof(mask));
+	arc4(&mask, sizeof(mask));
+	sprintf(buff, "x%lx", mask);
+	string = getenv(buff);
+#if DEBUGEXEC
+	fprintf(stderr, "getenv(%s)=%s\n", buff, string ? string : "<null>");
+#endif
+	l = strlen(buff);
+	if (!string) {
+		/* 1st */
+		sprintf(&buff[l], "=%lu %d", mask, argc);
+		putenv(strdup(buff));
+		return 0;
+	}
+	c = sscanf(string, "%lu %d%c", &m, &a, buff);
+	if (c == 2 && m == mask) {
+		/* 3rd */
+		rmarg(environ, &string[-l - 1]);
+		return 1 + (argc - a);
+	}
+	return -1;
+}
+
+void chkenv_end(void){}
+
+#if HARDENING
+
+static void gets_process_name(const pid_t pid, char * name) {
+	char procfile[BUFSIZ];
+	sprintf(procfile, "/proc/%d/cmdline", pid);
+	FILE* f = fopen(procfile, "r");
+	if (f) {
+		size_t size;
+		size = fread(name, sizeof (char), sizeof (procfile), f);
+		if (size > 0) {
+			if ('\n' == name[size - 1])
+				name[size - 1] = '\0';
+		}
+		fclose(f);
+	}
+}
+
+void hardening() {
+    prctl(PR_SET_DUMPABLE, 0);
+    prctl(PR_SET_PTRACER, -1);
+
+    int pid = getppid();
+    char name[256] = {0};
+    gets_process_name(pid, name);
+
+    if (   (strcmp(name, "bash") != 0) 
+        && (strcmp(name, "/bin/bash") != 0) 
+        && (strcmp(name, "sh") != 0) 
+        && (strcmp(name, "/bin/sh") != 0) 
+        && (strcmp(name, "sudo") != 0) 
+        && (strcmp(name, "/bin/sudo") != 0) 
+        && (strcmp(name, "/usr/bin/sudo") != 0)
+        && (strcmp(name, "gksudo") != 0) 
+        && (strcmp(name, "/bin/gksudo") != 0) 
+        && (strcmp(name, "/usr/bin/gksudo") != 0) 
+        && (strcmp(name, "kdesu") != 0) 
+        && (strcmp(name, "/bin/kdesu") != 0) 
+        && (strcmp(name, "/usr/bin/kdesu") != 0) 
+       )
+    {
+        printf("Operation not permitted\n");
+        kill(getpid(), SIGKILL);
+        exit(1);
+    }
+}
+
+#endif /* HARDENING */
+
+#if !TRACEABLE
+
+#define _LINUX_SOURCE_COMPAT
+#include <sys/ptrace.h>
+#include <sys/types.h>
+#include <sys/wait.h>
+#include <fcntl.h>
+#include <signal.h>
+#include <stdio.h>
+#include <unistd.h>
+
+#if !defined(PT_ATTACHEXC) /* New replacement for PT_ATTACH */
+   #if !defined(PTRACE_ATTACH) && defined(PT_ATTACH)
+       #define PT_ATTACHEXC	PT_ATTACH
+   #elif defined(PTRACE_ATTACH)
+       #define PT_ATTACHEXC PTRACE_ATTACH
+   #endif
+#endif
+
+void untraceable(char * argv0)
+{
+	char proc[80];
+	int pid, mine;
+
+	switch(pid = fork()) {
+	case  0:
+		pid = getppid();
+		/* For problematic SunOS ptrace */
+#if defined(__FreeBSD__)
+		sprintf(proc, "/proc/%d/mem", (int)pid);
+#else
+		sprintf(proc, "/proc/%d/as",  (int)pid);
+#endif
+		close(0);
+		mine = !open(proc, O_RDWR|O_EXCL);
+		if (!mine && errno != EBUSY)
+			mine = !ptrace(PT_ATTACHEXC, pid, 0, 0);
+		if (mine) {
+			kill(pid, SIGCONT);
+		} else {
+			perror(argv0);
+			kill(pid, SIGKILL);
+		}
+		_exit(mine);
+	case -1:
+		break;
+	default:
+		if (pid == waitpid(pid, 0, 0))
+			return;
+	}
+	perror(argv0);
+	_exit(1);
+}
+#endif /* !TRACEABLE */
+
+char * xsh(int argc, char ** argv)
+{
+	char * scrpt;
+	int ret, i, j;
+	char ** varg;
+	char * me = argv[0];
+	if (me == NULL) { me = getenv("_"); }
+	if (me == 0) { fprintf(stderr, "E: neither argv[0] nor $_ works."); exit(1); }
+
+	ret = chkenv(argc);
+	stte_0();
+	 key(pswd, pswd_z);
+	arc4(msg1, msg1_z);
+	arc4(date, date_z);
+	if (date[0] && (atoll(date)<time(NULL)))
+		return msg1;
+	arc4(shll, shll_z);
+	arc4(inlo, inlo_z);
+	arc4(xecc, xecc_z);
+	arc4(lsto, lsto_z);
+	arc4(tst1, tst1_z);
+	 key(tst1, tst1_z);
+	arc4(chk1, chk1_z);
+	if ((chk1_z != tst1_z) || memcmp(tst1, chk1, tst1_z))
+		return tst1;
+	arc4(msg2, msg2_z);
+	if (ret < 0)
+		return msg2;
+	varg = (char **)calloc(argc + 10, sizeof(char *));
+	if (!varg)
+		return 0;
+	if (ret) {
+		arc4(rlax, rlax_z);
+		if (!rlax[0] && key_with_file(shll))
+			return shll;
+		arc4(opts, opts_z);
+#if HARDENING
+	    arc4_hardrun(text, text_z);
+	    exit(0);
+       /* Seccomp Sandboxing - Start */
+       seccomp_hardening();
+#endif
+		arc4(text, text_z);
+		arc4(tst2, tst2_z);
+		 key(tst2, tst2_z);
+		arc4(chk2, chk2_z);
+		if ((chk2_z != tst2_z) || memcmp(tst2, chk2, tst2_z))
+			return tst2;
+		/* Prepend hide_z spaces to script text to hide it. */
+		scrpt = malloc(hide_z + text_z);
+		if (!scrpt)
+			return 0;
+		memset(scrpt, (int) ' ', hide_z);
+		memcpy(&scrpt[hide_z], text, text_z);
+	} else {			/* Reexecute */
+		if (*xecc) {
+			scrpt = malloc(512);
+			if (!scrpt)
+				return 0;
+			sprintf(scrpt, xecc, me);
+		} else {
+			scrpt = me;
+		}
+	}
+	j = 0;
+#if BUSYBOXON
+	varg[j++] = "busybox";
+	varg[j++] = "sh";
+#else
+	varg[j++] = argv[0];		/* My own name at execution */
+#endif
+	if (ret && *opts)
+		varg[j++] = opts;	/* Options on 1st line of code */
+	if (*inlo)
+		varg[j++] = inlo;	/* Option introducing inline code */
+	varg[j++] = scrpt;		/* The script itself */
+	if (*lsto)
+		varg[j++] = lsto;	/* Option meaning last option */
+	i = (ret > 1) ? ret : 0;	/* Args numbering correction */
+	while (i < argc)
+		varg[j++] = argv[i++];	/* Main run-time arguments */
+	varg[j] = 0;			/* NULL terminated array */
+#if DEBUGEXEC
+	debugexec(shll, j, varg);
+#endif
+	execvp(shll, varg);
+	return shll;
+}
+
+int main(int argc, char ** argv)
+{
+#if SETUID
+   setuid(0);
+#endif
+#if DEBUGEXEC
+	debugexec("main", argc, argv);
+#endif
+#if HARDENING
+	hardening();
+#endif
+#if !TRACEABLE
+	untraceable(argv[0]);
+#endif
+	argv[1] = xsh(argc, argv);
+	fprintf(stderr, "%s%s%s: %s\n", argv[0],
+		errno ? ": " : "",
+		errno ? strerror(errno) : "",
+		argv[1] ? argv[1] : "<null>"
+	);
+	return 1;
+}